Refactor sharing to album
