+++ mobile/lib/modules/album/models/album_viewer_page_state.model.dart
