+++ mobile/lib/modules/album/models/asset_selection_page_result.model.dart
