+++ mobile/lib/modules/album/models/asset_selection_state.model.dart
