+++ mobile/lib/modules/album/providers/album_title.provider.dart
