+++ mobile/lib/modules/album/ui/album_action_outlined_button.dart
