+++ mobile/lib/modules/album/ui/album_viewer_editable_title.dart
@@ -2,7 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:immich_mobile/modules/sharing/providers/album_viewer.provider.dart';
+import 'package:immich_mobile/modules/album/providers/album_viewer.provider.dart';
 import 'package:openapi/api.dart';
 
 class AlbumViewerEditableTitle extends HookConsumerWidget {
