+++ mobile/lib/modules/album/ui/shared_album_thumbnail_image.dart
