+++ mobile/lib/modules/album/ui/sharing_sliver_appbar.dart
