+++ mobile/lib/modules/album/views/library_page.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+class LibraryPage extends HookConsumerWidget {
+  const LibraryPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return const Scaffold(
+      body: Center(
+        child: Text('LibraryPage'),
+      ),
+    );
+  }
+}
